perf: update insetUtility to maintain current support
BREAKING CHANGE: This commit removes the legacy insetUtility mixin and
replaces it with a static output of the generated legacy values from the
previous version of WCSS and design tokens.

Users will not get a failure when importing the file reference
src/utilityMixins/_insetUtility.scss but instead of the mixin generating
a subset of selectors, all the static selectors will be added to the
project's CSS output.

This legacy feature has also been designated as DEPRECATED and should be
deleted with the next MAJOR release.