perf: update package config #171
BREAKING CHANGE: This commit updates the .npmignore package config to
reduce the scope of the packaged files and directories

This commit also removes the necessity of copying all the src files to a
dist dir and instead allows the src dir to be added to the package.
Consumers will have to update dir references from ./dist to ./src in
most import cases.

Processed files are still located in the ./dist dir.

Changes to be committed:
modified:   .github/workflows/testPublish.yml
modified:   .npmignore
modified:   README.md
modified:   package.json