fix: update utility import reference
This change is a response to an incorrectly referenced utility reference
from the auroElement styles. With the last release all the non-processed
files were removed from a ./dist dir and the package contains a ./src
dir for use. This move broke the previous utility ref.

Changes to be committed:
modified:   docs/index.html
modified:   src/auroElement/_auroElement.scss