perf: update utilities to support fallbacks #184
These fallbacks were intentionally not updates as the assumption was
that the user would choose either standard CSS or Sass and get the
expected return. This was proven false and the work was to add the
necessary fallbacks to be used regardless of CSS or Sass choice.

Changes to be committed:
modified:   src/utilityMixins/_anchor-roleButton.scss
modified:   src/utilityMixins/_anchor-roleTab.scss