perf: add duplicate .mjs file for SSR support
The additional auroElement.mjs file was added to aid in the effort to
support SSR development and Svelte.

This file was added as not to break any existing references to this file
that does not require SSR support. It is noted that the .js version is
deprecated and will be removed at a later date.

Changes to be committed:
new file:   src/auroElement/auroElement.mjs

Changes to be committed:
modified:   docs/index.html
modified:   src/auroElement/README.md
new file:   src/auroElement/auroElement.mjs