perf: remove unused breakpoint mixins #193
This update removes mixins that were not properly supported by the
design tokens that it required as Sass variables.

This is NOT a breaking change as, there are no tests to support these
mixins and, as stated, there are no variables to support the mixins
either, so anyone using this in the public space would not be able to
compile the Sass to CSS due to the missing variables.

This is the removal of code that should have not been added and is not
supported the Auro design guidelines.

Changes to be committed:
modified:   docs/index.html
modified:   src/_breakpoints.scss