docs: update content of README #190
This commit updates core content for how to install the necessary bits
of WCSS for optimal use in new projects.

Updates include placing priority emphasis on pre-processed CSS assets
and updating the install of npm resources to remove references to the
`~` deprecated shortcut.

On branch dsande/docs/190
Changes to be committed:
modified:   README.md